% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexeInvMC.R
\name{hexeInvMC}
\alias{hexeInvMC}
\title{hexeInvMC}
\usage{
hexeInvMC(
  x,
  ux,
  y,
  uy,
  y0,
  uy0,
  dist_x = "norm",
  dist_y = "norm",
  dist_y0 = "norm",
  n_sim = 10000
)
}
\arguments{
\item{x}{Vector numérico con los valores de concentración (x).}

\item{ux}{Vector numérico con las incertidumbres típicas de x.}

\item{y}{Vector numérico con los valores de respuesta (y).}

\item{uy}{Vector numérico con las incertidumbres típicas de y.}

\item{y0}{Valor observado a invertir.}

\item{uy0}{Incertidumbre típica de y0.}

\item{dist_x}{Distribución para x: "norm", "unif" o "triangle".}

\item{dist_y}{Distribución para y: "norm", "unif" o "triangle".}

\item{dist_y0}{Distribución para y0: "norm", "unif" o "triangle".}

\item{n_sim}{Número de simulaciones Monte Carlo. Por defecto 10000.}
}
\value{
(Invisiblemente) un objeto tipo lista con clase "hexeInvMC_result".
}
\description{
Regresión inversa con incertidumbre en ambos ejes mediante simulación Monte Carlo extendida.
Selección automática de modelo polinómico de grado 1 o 2 (ambos lineales en los parámetros),
diagnóstico y reporte metrológico con un gráfico principal.
}
