% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexeInvMC.R
\name{hexeInvMC}
\alias{hexeInvMC}
\title{hexeInvMC: Regresión inversa con incertidumbre}
\usage{
hexeInvMC(
  x,
  ux,
  y,
  uy,
  y0,
  uy0,
  n_sim = 10000,
  dist_x = "norm",
  dist_y = "norm",
  dist_y0 = "norm"
)
}
\arguments{
\item{x}{Vector numérico con valores de x (por ejemplo, concentración)}

\item{ux}{Vector numérico con incertidumbre estándar asociada a x}

\item{y}{Vector numérico con valores de y (por ejemplo, absorbancia)}

\item{uy}{Vector numérico con incertidumbre estándar asociada a y}

\item{y0}{Valor de respuesta instrumental observado para estimar su correspondiente x}

\item{uy0}{Incertidumbre estándar asociada a y0}

\item{n_sim}{Número de simulaciones (por defecto 10000)}

\item{dist_x}{Distribución para simular x: "norm", "unif", "triangle"}

\item{dist_y}{Distribución para simular y: "norm", "unif", "triangle"}

\item{dist_y0}{Distribución para simular y0: "norm", "unif", "triangle"}
}
\value{
Lista con el modelo seleccionado, el valor estimado de x y su incertidumbre combinada uc(x).
También imprime diagnóstico y genera una gráfica con ggplot2.
}
\description{
Realiza estimación inversa considerando incertidumbre en ambos ejes mediante simulación Monte Carlo extendida.
}
\examples{
x <- c(1.2, 1.9, 2.9, 4.0, 4.7, 5.9)
ux <- c(0.2, 0.2, 0.2, 0.2, 0.2, 0.2)
y <- c(3.4, 4.4, 7.2, 8.5, 10.8, 13.5)
uy <- c(0.2, 0.2, 0.2, 0.4, 0.4, 0.4)
hexeInvMC(x, ux, y, uy, y0 = 10.5, uy0 = 0.25)
}
